requirements.txt
Django==6.0.4
djangorestframework==3.15.2
drf-spectacular==0.28.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-modeltranslation==0.19.12
django-unfold==0.62.0
Pillow==11.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
